putchar() and single digits

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -5,12 +5,12 @@
 */
 int main(void)
 {
-long i;
+int i;
 for (i = 0; i < 10; i++)
 {
-if ((i == 0) || (i%10 + '0'))
+if ((i == 0) || (i % 10 + '0'))
 {
-putchar(('0') || (i%10 + '0'));
+putchar(('0') || (i % 10 + '0'));
 }
 }
 putchar('\n');
